requirements.txt
# Core quantum computing libraries
qiskit>=0.45.0
qiskit-aer>=0.13.0
qiskit-machine-learning>=0.7.0

# Machine learning and numerical computing
numpy>=1.24.0
scipy>=1.11.0
scikit-learn>=1.3.0

# Deep learning (optional, for hybrid models)
torch>=2.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Data handling
pandas>=2.0.0

# Progress bars and utilities
tqdm>=4.65.0

# Jupyter support
jupyter>=1.0.0
ipykernel>=6.25.0

# Testing
pytest>=7.4.0

# Configuration management
pyyaml>=6.0.0
